Add dev container & related changes
The initial implementation pulled the golang src from github master.
This makes it difficult to test local uncommited, unmerged changes. This
change removes the dev flag from the 'master' container and introduces a
new 'dev' container.

The dev container builds qmstr binaries from source mounted as a volume
in the container on runtime. So it's meant to be invoked for testing
local changes. See updated README.md for example.

Also a build.sh script was added to save some milliseconds :)